requirements.txt
flask
requests
pyaudio
elevenlabs
python-dotenv
ultralytics
anthropic
pillow
